Refactor Ui to package
